Ontology/ontology_enrichment_scripts/relationship_enricher/requirements.txt
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0
networkx>=3.0.0 